requirements.txt
torchsummary==1.5.1
loguru
opencv_python
tensorboard
tqdm